+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Players with boosted health healing when killing mobs
+
 ## [1.21.1-0.2.1] - 2025-01-31
 Blights are still pending a lot of work... Please be patient with me :)
 ### Added